+++ Src/MoneyFox.Ui/AppShell.xaml.cs
@@ -11,6 +11,7 @@ using Views.Categories.CategorySelection;
 using Views.Categories.ModifyCategory;
 using Views.Dashboard;
 using Views.Payments;
+using Views.Payments.PaymentList;
 using Views.Payments.PaymentModification;
 using Views.Settings;
 using Views.Setup;
