+++ Src/MoneyFox.Ui/InversionOfControl/MoneyFoxConfig.cs
@@ -19,6 +19,7 @@ using Views.Categories.ModifyCategory;
 using Views.Dashboard;
 using Views.OverflowMenu;
 using Views.Payments;
+using Views.Payments.PaymentList;
 using Views.Payments.PaymentModification;
 using Views.Popups;
 using Views.Settings;
